+++ packages/core-data/src/index.js
@@ -22,23 +22,29 @@ import { unlock } from './private-apis';
 // The "kind" and the "name" of the entity are combined to generate these shortcuts.
 
 const entitySelectors = rootEntitiesConfig.reduce( ( result, entity ) => {
-	const { kind, name } = entity;
+	const { kind, name, plural } = entity;
 	result[ getMethodName( kind, name ) ] = ( state, key, query ) =>
 		selectors.getEntityRecord( state, kind, name, key, query );
-	result[ getMethodName( kind, name, 'get', true ) ] = ( state, query ) =>
-		selectors.getEntityRecords( state, kind, name, query );
+
+	if ( plural ) {
+		result[ getMethodName( kind, plural, 'get' ) ] = ( state, query ) =>
+			selectors.getEntityRecords( state, kind, name, query );
+	}
 	return result;
 }, {} );
 
 const entityResolvers = rootEntitiesConfig.reduce( ( result, entity ) => {
-	const { kind, name } = entity;
+	const { kind, name, plural } = entity;
 	result[ getMethodName( kind, name ) ] = ( key, query ) =>
 		resolvers.getEntityRecord( kind, name, key, query );
-	const pluralMethodName = getMethodName( kind, name, 'get', true );
-	result[ pluralMethodName ] = ( ...args ) =>
-		resolvers.getEntityRecords( kind, name, ...args );
-	result[ pluralMethodName ].shouldInvalidate = ( action ) =>
-		resolvers.getEntityRecords.shouldInvalidate( action, kind, name );
+
+	if ( plural ) {
+		const pluralMethodName = getMethodName( kind, plural, 'get' );
+		result[ pluralMethodName ] = ( ...args ) =>
+			resolvers.getEntityRecords( kind, name, ...args );
+		result[ pluralMethodName ].shouldInvalidate = ( action ) =>
+			resolvers.getEntityRecords.shouldInvalidate( action, kind, name );
+	}
 	return result;
 }, {} );